backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain & AI
langchain>=0.1.10
langchain-google-genai>=0.0.9
langgraph>=0.0.26
google-generativeai>=0.4.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery & Task Queue
celery==5.3.6
celery[redis]==5.3.6
kombu==5.3.5

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0
python-socketio==5.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Email

jinja2==3.1.3

# Monitoring & Logging
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
